medease-bd/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
pydantic
llama-cpp-python
sentence-transformers
numpy
pandas
httpx
redis
psutil
